% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CinicalTrialInterimAnalysis.R
\name{CinicalTrialInterimAnalysis}
\alias{CinicalTrialInterimAnalysis}
\title{Clinical Trial Interim Report}
\usage{
CinicalTrialInterimAnalysis()
}
\description{
This function will run a Shiny application
}
