% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinicalTrialInterimAnalysis.R
\name{ClinicalTrialInterimAnalysis}
\alias{ClinicalTrialInterimAnalysis}
\title{Clinical Trial Interim Report}
\usage{
ClinicalTrialInterimAnalysis()
}
\description{
This function will run a Shiny application
}
