% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinicalTrialInterimReport.R
\name{ClinicalTrialInterimReport}
\alias{ClinicalTrialInterimReport}
\title{Clinical Trial Interim Report}
\usage{
ClinicalTrialInterimReport()
}
\description{
This function will run a Shiny application
}
